+++ packages/dumb/src/index.ts
@@ -66,6 +66,7 @@ import { PrintView } from './modules/print/print-view/print-view'
 import { LayoutColumnType } from './models/enums/layout-column-type.enum'
 import { buildLayoutColumns } from './modules/print/print-view/print-view.util'
 import { getPatientFullName } from './utils/patient/patient.util'
+import { formatBirthdate } from './utils/datetime/datetime.util'
 
 export {
   AverageDailyDoseStat,
@@ -82,6 +83,7 @@ export {
   ConfidentialTooltip,
   FocusedRangeLabels,
   FoodTooltip,
+  formatBirthdate,
   getPatientFullName,
   HistoryParameterTable,
   formatBgValue,