+++ packages/yourloops/components/patient-list/column-selector-popover.tsx
@@ -69,22 +69,34 @@ export const ColumnSelectorPopover: FunctionComponent<ColumnSelectorPopoverProps
       tooltip: t('un-removable-column')
     },
     {
-      name: PatientListColumns.MonitoringAlerts,
-      checked: updatedColumnsModel[PatientListColumns.MonitoringAlerts],
-      hcpOnly: true
+      name: PatientListColumns.Age,
+      checked: updatedColumnsModel[PatientListColumns.Age]
+    },
+    {
+      name: PatientListColumns.DateOfBirth,
+      checked: updatedColumnsModel[PatientListColumns.DateOfBirth]
+    },
+    {
+      name: PatientListColumns.Gender,
+      checked: updatedColumnsModel[PatientListColumns.Gender]
     },
     {
       name: PatientListColumns.System,
       checked: updatedColumnsModel[PatientListColumns.System]
     },
     {
-      name: PatientListColumns.LastDataUpdate,
-      checked: updatedColumnsModel[PatientListColumns.LastDataUpdate]
+      name: PatientListColumns.MonitoringAlerts,
+      checked: updatedColumnsModel[PatientListColumns.MonitoringAlerts],
+      hcpOnly: true
     },
     {
       name: PatientListColumns.Messages,
       checked: updatedColumnsModel[PatientListColumns.Messages],
       hcpOnly: true
+    },
+    {
+      name: PatientListColumns.LastDataUpdate,
+      checked: updatedColumnsModel[PatientListColumns.LastDataUpdate]
     }
   ]