+++ packages/yourloops/components/patient-list/models/grid-row.model.ts
@@ -32,6 +32,9 @@ export interface GridRowModel {
   id: string
   [PatientListColumns.Flag]?: Patient
   [PatientListColumns.Patient]: Patient
+  [PatientListColumns.DateOfBirth]?: Patient
+  [PatientListColumns.Age]?: number
+  [PatientListColumns.Gender]?: string
   [PatientListColumns.System]?: string
   [PatientListColumns.MonitoringAlerts]?: Patient
   [PatientListColumns.LastDataUpdate]?: string