+++ packages/yourloops/lib/auth/models/enums/gender.enum.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2023, Diabeloop
+ * Copyright (c) 2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,28 +25,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { screen, waitFor } from '@testing-library/react'
-import { mockPatientLogin } from '../../mock/patient-login.mock'
-import { checkPatientNavBarAsPatient } from '../../assert/patient-nav-bar.assert'
-import { minimalTrendViewData, mockDataAPI } from '../../mock/data.api.mock'
-import { renderPage } from '../../utils/render'
-import { checkPatientLayout } from '../../assert/layout.assert'
-import { unmonitoredPatientAsTeamMember } from '../../data/patient.api.data'
-
-describe('Trends view for patient', () => {
-  beforeAll(() => {
-    mockPatientLogin(unmonitoredPatientAsTeamMember)
-  })
-
-  it('should render correct layout', async () => {
-    mockDataAPI(minimalTrendViewData)
-    const router = renderPage('/trends')
-    await waitFor(() => {
-      expect(router.state.location.pathname).toEqual('/trends')
-    })
-
-    expect(await screen.findByTestId('patient-nav-bar', {}, { timeout: 3000 })).toBeVisible()
-    checkPatientNavBarAsPatient()
-    await checkPatientLayout(`${unmonitoredPatientAsTeamMember.profile.firstName} ${unmonitoredPatientAsTeamMember.profile.lastName}`)
-  })
-})
+export enum Gender {
+  Female = 'F',
+  Indeterminate = 'I',
+  Male = 'M',
+  NotDefined = ''
+}