+++ CHANGELOG.md
@@ -28,6 +28,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2231 Implement new medical reports design for patients
 - YLP-2265 Disable Add patient button when private practice is selected
 - YLP-2322 Pending patients should not display glycemia specific data
+- YLP-2323 Add new columns from existing data in patient list
 
 ### Engineering use
 - YLP-2044 Remove redux settings store